error managed. starting hardening.

+++ xpeppers.sh
@@ -28,8 +28,19 @@ fi
 
 echo "Downloading Wordpress version ${version}"
 wget -c https://wordpress.org/wordpress-${version}.tar.gz -P /tmp
+if [[ "$?" -ne 0 ]]
+then
+	echo "Errors getting wordpress ${version}. Aborting."
+	exit 1
+fi 
+
 echo "Downloading md5"
 wget -c https://wordpress.org/wordpress-${version}.tar.gz.md5 -P /tmp
+if [[ "$?" -ne 0 ]]
+then
+	echo "Errors getting wordpress ${version} MD5. Aborting."
+	exit 1
+fi
 
 echo "Calculating md5 for Wordpress"
 md5=$(md5sum /tmp/wordpress-${version}.tar.gz | awk '{print $1}')
@@ -37,13 +48,18 @@ echo "Comparing md5"
 if [[ $(cat /tmp/wordpress-${version}.tar.gz.md5) != "$md5" ]]
 then
 	echo "Md5 are different. Aborting."
-	exit 0
+	exit 1
 else
 	echo "Md5 check ok"
 fi
 
 echo "Installing wordpress prerequisites"
 sudo apt-get install apache2 mysql-server php5 php5-mysqlnd-ms
+if [[ "$?" -ne 0 ]]
+then
+	echo "Errors getting wordpress prerequisites"
+	exit 1
+fi
 
 echo "Setting up mysql"
 cp mysql_template.sql mysql.sql
@@ -52,6 +68,11 @@ sed -i -e "s/user_name_here/$mysqluser/g" mysql.sql
 sed -i -e "s/password_here/$mysqlpasswd/g" mysql.sql
 
 cat mysql.sql | mysql -u root -p
+if [[ "$?" -ne 0 ]]
+then
+	echo "Errors configuring MySQL. Aborting."
+	exit 1
+fi
 rm ./mysql.sql
 
 echo "Extracting wordpress"
@@ -78,6 +99,11 @@ then
 fi
 
 sudo service apache2 restart
+if [[ "$?" -ne 0 ]]
+then
+	echo "Errors restarting apache2. Aborting."
+	exit 1
+fi
 
 echo "Installation OK. Launching Firefox..."
